refactor: siteUser 관련 Type 클래스 패키지 이동
- PreparationStatus, Role을 siteuser.domain 로
