+++ src/main/java/com/example/solidconnection/auth/dto/SignUpRequest.java
@@ -2,8 +2,8 @@ package com.example.solidconnection.auth.dto;
 
 import com.example.solidconnection.siteuser.domain.AuthType;
 import com.example.solidconnection.siteuser.domain.SiteUser;
-import com.example.solidconnection.type.PreparationStatus;
-import com.example.solidconnection.type.Role;
+import com.example.solidconnection.siteuser.domain.PreparationStatus;
+import com.example.solidconnection.siteuser.domain.Role;
 import jakarta.validation.constraints.NotBlank;
 
 import java.util.List;
